Add HA files
